requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Optional: install separately if needed
# xgboost>=1.7.0

# Experiment Tracking
mlflow>=2.9.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Config & Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Data / Reports
matplotlib>=3.7.0
seaborn>=0.12.0
